refacto + add logscreen
