+++ interface/kv/log.kv
@@ -1,0 +1,61 @@
+<LogScreen>:
+    canvas.before:
+        Color:
+            rgba: 1, 1, 1, 1  # Fond blanc
+        Rectangle:
+            pos: self.pos
+            size: self.size
+    
+    BoxLayout:
+        orientation: 'vertical'
+        padding: 10
+        spacing: 10
+        
+        # En-tête
+        Label:
+            text: 'Logs'
+            size_hint_y: None
+            height: 40
+            font_size: '20sp'
+            color: 0, 0, 0, 1
+            bold: True
+        
+        # Zone de logs avec scroll
+        ScrollView:
+            id: log_scroll
+            do_scroll_x: False
+            bar_width: 10
+            bar_color: 0.5, 0.5, 0.5, 1
+            
+            Label:
+                id: log_label
+                text: root.log_text
+                size_hint_y: None
+                height: self.texture_size[1]
+                text_size: self.width, None
+                color: 0, 0, 0, 1
+                font_name: 'RobotoMono-Regular'
+                font_size: '12sp'
+                halign: 'left'
+                valign: 'top'
+                padding: 10, 10
+                markup: True
+        
+        # Boutons d'action
+        BoxLayout:
+            orientation: 'horizontal'
+            size_hint_y: None
+            height: 50
+            spacing: 10
+            
+            Button:
+                text: 'Effacer les logs'
+                background_color: 0.9, 0.9, 0.9, 1
+                color: 0, 0, 0, 1
+                on_release: root.clear_logs()
+            
+            Button:
+                text: 'Retour'
+                background_color: 0.3, 0.5, 0.9, 1
+                color: 1, 1, 1, 1
+                on_release: root.go_back()
