+++ interface/kv/main.kv
@@ -5,29 +5,27 @@
         Rectangle:
             pos: self.pos
             size: self.size
-
     BoxLayout:
         orientation: "vertical"
         padding: 10
         spacing: 10
-
         BoxLayout:
             id: main_container
             orientation: "vertical"
-
         BoxLayout:
             size_hint_y: None
             height: 50
             spacing: 10
-
             Button:
                 text: "Default"
                 on_release: root.load_defaults()
-
             Button:
                 text: "Reset tout"
                 on_release: root.reset_all()
-
             Button:
                 text: "Sauvegarder"
                 on_release: root.save_elements()
+            Button:
+                text: "Exécuter"
+                background_color: 0.2, 0.7, 0.3, 1
+                on_release: root.execute_script()
