+++ interface/screens/__init__.py
@@ -1,0 +1,5 @@
+from .config_screen import ConfigScreen
+from .main_screen import MainScreen
+from .log_screen import LogScreen
+
+__all__ = ['ConfigScreen', 'MainScreen', 'LogScreen']
